cipherguard/key_management.py
from cryptography.fernet import Fernet
from cryptography.hazmat.primitives.asymmetric import rsa
from cryptography.hazmat.primitives import serialization

def generate_key(algorithm):
    if algorithm == 'AES':
        return Fernet.generate_key().decode()
    elif algorithm == 'RSA':
        private_key = rsa.generate_private_key(
            public_exponent=65537,
            key_size=2048
        )
        public_key = private_key.public_key()
        pem = public_key.public_bytes(
            encoding=serialization.Encoding.PEM,
            format=serialization.PublicFormat.SubjectPublicKeyInfo
        )
        return pem.decode()

def store_key(key, algorithm):
    # Implement secure key storage logic here
    pass

def retrieve_key(key_id, algorithm):
    # Implement key retrieval logic here
    pass

# In app/key_management.py

import os
import json

KEY_STORE = {}  # In-memory key store. In a real app, use a secure database.

def store_key(key, algorithm):
    key_id = os.urandom(16).hex()  # Generate a random key ID
    KEY_STORE[key_id] = {'key': key, 'algorithm': algorithm}
    return key_id

def retrieve_key(key_id, algorithm):
    key_data = KEY_STORE.get(key_id)
    if key_data and key_data['algorithm'] == algorithm:
        return key_data['key']
    return None